Unify Background Effects Layout
