+++ frontend/src/components/BackgroundEffects/BackgroundEffectsLayout/index.tsx
