--- frontend/src/components/WaitingRoom/BackgroundEffects/BackgroundEffectsLayout/Index.tsx
@@ -1,3 +1,0 @@
-import BackgroundEffectsLayout from './BackgroundEffectsLayout';
-
-export default BackgroundEffectsLayout;
